requirements.txt
discord.py
pynacl
requests
dotenv
aiosqlite